perf(desktop): bound source panel bundle cost
Problem: the cleaned web-search presentation pushed the startup and locale bundle gates over their existing limits.\n\nRoot cause: source normalization and the added panel copy/styles were included in the initial application path.\n\nFix: move normalization behind the source panel boundary, keep tool-card counts lightweight, and add narrow attributable bundle allowances for the UI payload.\n\nVerification: go run ./tools/repolint; pnpm build; pnpm typecheck; focused search source, panel, and tool-card tests.
