+++ desktop/frontend/src/locales/en.ts
@@ -2919,9 +2919,7 @@ export const en = {
   "msg.imageAttachment": "Image",
   "msg.workspaceReference": "Workspace reference",
   "sources.title": "Sources",
-  "sources.count": "{n}",
   "sources.hidden": "{n} hidden",
-  "sources.noValid": "No valid sources",
   "sources.open": "Open source",
   "sources.copyLink": "Copy link",
   "msg.folderReference": "Folder · Workspace reference",
