+++ desktop/frontend/src/locales/zh-TW.ts
@@ -2019,9 +2019,7 @@ export const zhTW: Record<DictKey, string> = {
   "msg.imageAttachment": "圖片",
   "msg.workspaceReference": "工作區引用",
   "sources.title": "來源",
-  "sources.count": "{n} 條",
   "sources.hidden": "已隱藏 {n} 條",
-  "sources.noValid": "沒有有效來源",
   "sources.open": "開啟來源",
   "sources.copyLink": "複製連結",
   "msg.folderReference": "資料夾 · 工作區引用",
