+++ desktop/frontend/src/locales/zh.ts
@@ -2922,9 +2922,7 @@ export const zh: Record<DictKey, string> = {
   "msg.imageAttachment": "图片",
   "msg.workspaceReference": "工作区引用",
   "sources.title": "来源",
-  "sources.count": "{n} 条",
   "sources.hidden": "已隐藏 {n} 条",
-  "sources.noValid": "没有有效来源",
   "sources.open": "打开来源",
   "sources.copyLink": "复制链接",
   "msg.folderReference": "文件夹 · 工作区引用",
